+++ docs/changelog.md
@@ -21,6 +21,10 @@ Added:
 
 - The rendering of explorer tiles can now be fully customized. For each kind of special tile (e.g. visited, missing, new in this activity), you can set a fill color, border color, border style and stripe color. ([GH-366](https://github.com/martin-ueding/geo-activity-playground/issues/366))
 
+Changed:
+
+- Enable “Mapterhorn Hillshade” overlay by default. ([GH-489](https://github.com/martin-ueding/geo-activity-playground/issues/489))
+
 
 ## Version 1.46.0 — 2026-08-03